Change user list API to be course-specific
