+++ analytics_data_api/v0/views/users.py
@@ -1,52 +1,10 @@
 """
 API methods for user data.
 """
-from django.http import Http404
 from rest_framework import generics
 
 from analytics_data_api.v0.models import UserProfile
 from analytics_data_api.v0.serializers import UserProfileSerializer
-
-
-class UserListView(generics.ListAPIView):
-    """
-    Get the (paginated) list of all users (students and staff).
-
-    **Example Request**
-
-        GET /api/v0/users/
-
-    **Additional parameters**
-
-        These query parameters can be specified to control pagination of the results:
-
-            * limit: Specify how many results to return per page (default is 100).
-            * page: Specify which page of results. 1 is the first page.
-
-    **Response Values**
-
-        Returns an object with pagination info and a 'results' entry that is a collection of user
-        objects. Each user object contains:
-
-            * id: The user's ID (integer)
-            * username: The username (string)
-            * last_login: When the user last logged in to the LMS/Studio (datetime)
-            * date_joined: When the user registered (datetime)
-            * is_staff: True if the user is staff (boolean)
-            * email: The user's email address (string)
-            * name: The user's full name (string)
-            * gender: One of "male", "female", "other", or "unknown" (string)
-            * year_of_birth: Year of birth as integer or null
-            * level_of_education: String indicating self-reported education level, or "unknown"
-    """
-
-    serializer_class = UserProfileSerializer
-    paginate_by = 100  # When django-rest-framework is updated, convert this to LimitOffsetPagination
-    paginate_by_param = 'limit'
-
-    def get_queryset(self):
-        """Select the view count for a specific module"""
-        return UserProfile.objects.all()
 
 
 class UserProfileView(generics.RetrieveAPIView):
@@ -55,7 +13,7 @@ class UserProfileView(generics.RetrieveAPIView):
 
     **Example Request**
 
-        GET /api/v0/users/{user_id}/
+        GET /api/v0/users/{username}/
 
     **Response Values**
 
@@ -74,12 +32,6 @@ class UserProfileView(generics.RetrieveAPIView):
     """
 
     serializer_class = UserProfileSerializer
-    lookup_url_kwarg = 'user_id'
-
-    def get_queryset(self):
-        """Select the profile of a specific user"""
-        try:
-            user_id = int(self.kwargs.get('user_id'))
-        except ValueError:
-            raise Http404
-        return UserProfile.objects.filter(pk=user_id)
+    model = UserProfile
+    lookup_url_kwarg = 'username'
+    lookup_field = 'username'
